resolve nav nodes
