+++ core/src/luigi-config/communication.js
@@ -1,7 +1,15 @@
+import { relogin, getToken } from './navigation/navigation-helpers';
+
 export const communication = {
   customMessagesListeners: {
     'console.refreshNavigation': () => {
-      Luigi.configChanged('navigation.nodes');
+      const token = getToken()
+      if(token){
+        Luigi.configChanged('navigation.nodes');
+      }
+      else {
+        relogin()
+      }
     }
   }
 };
